+++ src/Lightbox.js
@@ -1,7 +1,7 @@
 import React, { Component, Fragment } from 'react'
 import PropTypes from 'prop-types'
 import noScroll from 'no-scroll'
-import LightboxMarkup from './LightboxMarkup'
+import Markup from './Markup'
 
 const keycodes = {
   esc: 27,
@@ -11,9 +11,6 @@ const keycodes = {
 
 class Lightbox extends Component {
   state = {
-    width: typeof window !== 'undefined' ? window.innerWidth : 0,
-    height: typeof window !== 'undefined' ? window.innerHeight : 0,
-    mouseIdle: false,
     showPortal: !!this.props.open,
     index: 0
   }
@@ -27,7 +24,6 @@ class Lightbox extends Component {
   componentWillReceiveProps(nextProps) {
     if (!this.props.open && nextProps.open) {
       this.handleOpen()
-      this.setState({ showPortal: true })
     }
     if (this.props.open && !nextProps.open) {
       this.handleClose()
@@ -40,32 +36,40 @@ class Lightbox extends Component {
     }
   }
 
-  triggerOpen = () => {
+  toggleOpen = () => {
     this.setState({
       showPortal: !this.state.showPortal
     })
   }
 
+  open = () => {
+    this.setState({
+      showPortal: true
+    })
+  }
+
+  close = () => {
+    this.setState({
+      showPortal: true
+    })
+  }
+
   handleOpen = () => {
     document.addEventListener('keydown', this.handleKeydown)
-    window.addEventListener('resize', this.handleWindowResize)
     document
       .querySelector('*')
       .addEventListener('mousemove', this.handleMousemove)
     noScroll.on()
+    this.open()
   }
 
   handleClose = () => {
     document.removeEventListener('keydown', this.handleKeydown)
-    window.removeEventListener('resize', this.handleWindowResize)
     document
       .querySelector('*')
       .removeEventListener('mousemove', this.handleMousemove)
     noScroll.off()
-  }
-
-  handleWindowResize = () => {
-    this.setState({ width: window.innerWidth, height: window.innerHeight })
+    this.close()
   }
 
   handleKeydown = e => {
@@ -75,38 +79,23 @@ class Lightbox extends Component {
       this.handleClickNext()
     } else if (e.keyCode === keycodes.esc && this.props.closeOnEsc) {
       this.handleClose()
-      this.props.onClose()
     }
-  }
-
-  handleMousemove = () => {
-    // Hide the actions buttons when move do not move for x seconds
-    clearTimeout(this.timeoutMouseIdle)
-    if (this.state.mouseIdle === true) {
-      this.setState({ mouseIdle: false })
-    }
-    this.timeoutMouseIdle = setTimeout(() => {
-      this.setState({ mouseIdle: true })
-    }, this.props.mouseIdleTimeout)
   }
 
   handleClickPrev = () => {
     if (this.state.index !== 0) {
       this.setState({ index: this.state.index - 1 })
-      this.props.onClickPrev()
     }
   }
 
   handleClickNext = () => {
     if (this.props.src[this.state.index + 1]) {
       this.setState({ index: this.state.index + 1 })
-      this.props.onClickNext()
     }
   }
 
   handleClickCloseArrow = () => {
-    this.triggerOpen()
-    this.props.onClose()
+    this.handleClose()
   }
 
   handleExited = () => {
@@ -116,69 +105,38 @@ class Lightbox extends Component {
   render () {
     const markupProps = {
       ...this.state,
-      transitionDuration: this.props.transitionDuration,
-      transitionStyles: this.props.transitionStyles,
-      triggerOpen: this.triggerOpen,
+      toggleOpen: this.toggleOpen,
       src: this.props.src
     }
 
     const childProps = {
       ...this.state,
-      triggerOpen: this.triggerOpen
+      toggleOpen: this.toggleOpen
     }
 
     const { children } = this.props
 
     return (
       <Fragment>
-        <LightboxMarkup {...markupProps} />
+        <Markup {...markupProps} />
         {children(childProps)}
       </Fragment>
     )
   }
 
   static propTypes = {
-    // Control if Lightbox is open or not
     open: PropTypes.bool,
-    // An array of image urls
     src: PropTypes.arrayOf(
       PropTypes.string
     ).isRequired,
-    // Is closable when user press esc key
     closeOnEsc: PropTypes.bool,
-    // Enable left and right arrow navigation
     keyboardNavigation: PropTypes.bool,
-    // The duration of the transition, in milliseconds see [react-transition-group docs](https://reactcommunity.org/react-transition-group/#Transition-prop-timeout)
-    transitionDuration: PropTypes.number,
-    // eslint-disable-next-line
-    // The animation object see [react-transition-group docs](https://reactcommunity.org/react-transition-group/#Transition)
-    transitionStyles: PropTypes.object,
-    // Timeout before hidding the actions buttons when mouse do not move (milliseconds)
-    mouseIdleTimeout: PropTypes.number,
-    // Function called when the previous image is requested
-    onClickPrev: PropTypes.func,
-    // Function called when the next image is requested
-    onClickNext: PropTypes.func,
-    // Function called when GooglePhoto is requested to be closed
-    onClose: PropTypes.func,
-    // HOC Usage
     children: PropTypes.func.isRequired
   }
 
   static defaultProps = {
     closeOnEsc: true,
-    keyboardNavigation: true,
-    mouseIdleTimeout: 5000,
-    transitionDuration: 200,
-    transitionStyles: {
-      default: {
-        transition: `opacity 200ms ease-in-out`,
-        opacity: 0
-      },
-      entering: { opacity: 0 },
-      entered: { opacity: 1 },
-      exiting: { opacity: 0 }
-    }
+    keyboardNavigation: true
   }
 }
